RF: `metadata()` is now a plain read/query command
Any manipulation functionality has been removed. The lost functionality
should come back via 'metadata backends' that may support manipulation,
as suggested by @yarikoptic in gh-2087.

However, it is extremely unlikely that all backends would use the same
interface (i.e. simple key-value stores vs graphs), hence it makes no
sense to preserve a git-annex style main interface here. Removal is
actually a big relief, due to the amount of semi-broken code that is
gone now.

- gh-2098 is no longer an issue
- fewer ties to datalad core (gh-2082)
- gh-2034 is no longer an issue, as we are not limited to what is
  supported in git-annex anymore
- gh-1972 is a non-issue because we no longer manipulate metadata
  (although some functionality has to come back in some form)
- gh-1941 is a non-issue as merges are gone now
