+++ datalad/metadata/tests/test_aggregation.py
@@ -10,30 +10,25 @@
 """Test metadata manipulation"""
 
 
-import os
 from os.path import join as opj
-from os.path import exists
 
 from datalad.api import metadata
 from datalad.distribution.dataset import Dataset
-from datalad.support.gitrepo import GitRepo
-from datalad.support.annexrepo import AnnexRepo
 
-from datalad.utils import chpwd
 
-from datalad.tests.utils import create_tree
 from datalad.tests.utils import with_tree
-from datalad.tests.utils import with_tempfile
-from datalad.tests.utils import assert_status
-from datalad.tests.utils import assert_in
-from datalad.tests.utils import assert_not_in
-from datalad.tests.utils import assert_raises
 from datalad.tests.utils import assert_result_count
 from datalad.tests.utils import assert_dict_equal
 from datalad.tests.utils import eq_
 from datalad.tests.utils import ok_clean_git
-from datalad.tests.utils import swallow_outputs
 from datalad.tests.utils import skip_direct_mode
+
+
+def _assert_metadata_empty(meta):
+    ignore = set(['@id', '@context'])
+    assert (not len(meta) or set(meta.keys()) == ignore), \
+        'metadata record is not empty: {}'.format(
+            {k: meta[k] for k in meta if k not in ignore})
 
 
 _dataset_hierarchy_template = {
@@ -60,7 +55,7 @@ def test_basic_aggregate(path):
     # TODO give datasets some more metadata to actually aggregate stuff
     base = Dataset(opj(path, 'origin')).create(force=True)
     sub = base.create('sub', force=True)
-    base.metadata(sub.path, init=dict(homepage='this'), apply2global=True)
+    #base.metadata(sub.path, init=dict(homepage='this'), apply2global=True)
     subsub = base.create(opj('sub', 'subsub'), force=True)
     base.add('.', recursive=True)
     ok_clean_git(base.path)
@@ -85,3 +80,49 @@ def test_basic_aggregate(path):
     agg_meta = base.metadata(recursive=True, return_type='list')
     for d, a in zip(direct_meta, agg_meta):
         assert_dict_equal(d, a)
+
+
+# tree puts aggregate metadata structures on two levels inside a dataset
+@with_tree(tree={
+    '.datalad': {
+        'metadata': {
+            'objects': {
+                'someshasum': '{"homepage": "http://top.example.com"}'},
+            'aggregate_v1.json': """\
+{
+    "sub/deep/some": {
+        "dataset_info": "objects/someshasum"
+    }
+}
+"""}},
+    'sub': {
+        '.datalad': {
+            'metadata': {
+                'objects': {
+                    'someotherhash': '{"homepage": "http://sub.example.com"}'},
+                'aggregate_v1.json': """\
+{
+    "deep/some": {
+        "dataset_info": "objects/someotherhash"
+    }
+}
+"""}}},
+})
+def test_aggregate_query(path):
+    ds = Dataset(path).create(force=True)
+    # no magic change to actual dataset metadata due to presence of
+    # aggregated metadata
+    res = ds.metadata(reporton='datasets')
+    assert_result_count(res, 1)
+    _assert_metadata_empty(res[0]['metadata'])
+    # but we can now ask for metadata of stuff that is unknown on disk
+    res = ds.metadata(opj('sub', 'deep', 'some'), reporton='datasets')
+    assert_result_count(res, 1)
+    eq_({'homepage': 'http://top.example.com'}, res[0]['metadata'])
+    # and the command will report the aggregated metadata as it is recorded
+    # in the dataset that is the closest parent on disk
+    ds.create('sub', force=True)
+    # same call as above, different result!
+    res = ds.metadata(opj('sub', 'deep', 'some'), reporton='datasets')
+    assert_result_count(res, 1)
+    eq_({'homepage': 'http://sub.example.com'}, res[0]['metadata'])
